% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\name{nils_estimate}
\alias{nils_estimate}
\title{Estimates according to Nils design}
\usage{
nils_estimate(psus, categories, tracts, values, tract_area, frame_area)
}
\description{
Estimates according to Nils design
}
