% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nils.R
\name{NilsEstimate}
\alias{NilsEstimate}
\alias{NilsEstimateBalanced}
\title{Estimate totals using the NILS hierarchical design (DEPRECATED)}
\usage{
NilsEstimate(
  plot_data,
  tract_data,
  psus,
  category_psu_map,
  area = 46519242.1175867,
  tract_area = 196 * 100 * pi
)

NilsEstimateBalanced(
  plot_data,
  tract_data,
  auxiliaries,
  psus,
  category_psu_map,
  area = 46519242.1175867,
  tract_area = 196 * 100 * pi,
  size_of_neighbourhood = 4L
)
}
\arguments{
\item{plot_data}{See \code{\link[=nilsier]{nilsier()}} \code{data}.}

\item{tract_data}{See \code{\link[=nilsier]{nilsier()}} \code{tracts}.}

\item{psus}{See \code{\link[=nilsier]{nilsier()}} \code{psus}.}

\item{category_psu_map}{See \code{\link[=nilsier]{nilsier()}} \code{categories}.}

\item{area}{See \code{\link[=nilsier]{nilsier()}} \code{frame_area}.}

\item{tract_area}{See \code{\link[=nilsier]{nilsier()}} \code{tract_area}.}

\item{auxiliaries}{See \code{\link[=nilsier]{nilsier()}} \code{tracts}}

\item{size_of_neighbourhood}{See \code{\link[=nilsier]{nilsier()}} \code{neighbourhood_size}.}
}
\value{
See return value of \code{\link[=nilsier]{nilsier()}}.
}
\description{
Deprecated in 0.2.0. Use \code{\link[=nilsier]{nilsier()}} instead.
}
\keyword{deprecated}
