% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_area}
\alias{prepare_area}
\title{Prepare an area}
\usage{
prepare_area(area)
}
\arguments{
\item{area}{An area value}
}
\value{
The value
}
\description{
Prepare an area value, or rejects it if non-positive
}
\examples{
prepped_area = prepare_area(2.0);

}
\seealso{
Other prepare:
\code{\link[=prepare_categories]{prepare_categories()}},
\code{\link[=prepare_data]{prepare_data()}},
\code{\link[=prepare_psus]{prepare_psus()}},
\code{\link[=prepare_tracts]{prepare_tracts()}}
}
\concept{prepare}
