% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_data}
\alias{prepare_data}
\alias{PreparePlotData}
\title{Prepare a tract data object}
\usage{
prepare_data(
  data,
  tract = ~tract,
  category = ~category,
  dw = ~dw,
  value = ~value
)

PreparePlotData(data, tid, cat, dw, y)
}
\arguments{
\item{data}{A data frame (or list) containing the survey data:
\describe{
\item{tract}{the tract identifier (integer).}
\item{category}{the category identifier (integer).}
\item{dw}{the design weight of the data (float).}
\item{value}{the recorded value (float).}
}
If the columns (or items) are named, the default names can be overidden by the corresponding
argument.
If the columns (or items) are unnamed, the order of the columns (or items) is assumed to be
according to the order given by the parameter list.}

\item{tract}{A formula specifying the column (or item) containing tract IDs.}

\item{category}{A formula specifying the column (or item) containing category IDs.}

\item{dw}{A formula specifying the column (or item) containing design weights.}

\item{value}{A formula specifying the column (or item) containing the variable of interest.}
}
\value{
A data frame
\describe{
\item{tract}{the tract identifier (integer).}
\item{category}{the category identifier (integer).}
\item{dw}{the design weight of the data (float).}
\item{value}{the recorded value (float).}
}
}
\description{
Prepare a tract data object (or plot data), which can be either a list or a data frame.
}
\examples{
prepped_data = prepare_data(plots);

}
\seealso{
Other prepare:
\code{\link[=prepare_area]{prepare_area()}},
\code{\link[=prepare_categories]{prepare_categories()}},
\code{\link[=prepare_psus]{prepare_psus()}},
\code{\link[=prepare_tracts]{prepare_tracts()}}
}
\concept{prepare}
