% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_psus}
\alias{prepare_psus}
\title{Prepare a PSU object}
\usage{
prepare_psus(psus, tracts = NULL, neighbourhood_size = 4L)
}
\arguments{
\item{psus}{A vector of PSU identifiers (integer) in order from the largest to the smallest, or;
a data frame, list, matrix, or vector containing the PSUs:
\describe{
\item{psu}{the PSU identifiers (integer) in order from the largest to the smallest.}
\item{nn_size}{(optional) the neighbourhood size to use (integer > \code{1L}). Will be derived from
\code{neighbourhood_size} if not given.}
}
If the columns (or items) are unnamed, or if a matrix is supplied, the order of the columns
(or items) are assumed to follow the order given by the list above.
If a matrix is supplied, the order is also assumed to be in the parameter list order.}

\item{tracts}{A list or data.frame of tracts with the following items (columns):
\describe{
\item{psu}{the PSU ID of the tract.}
}}

\item{neighbourhood_size}{The default neighbourhood size to use for the smallest PSU
(integer > \code{1L}). The neighbourhood size for larger PSUs scales linearly.}
}
\value{
A data frame
\describe{
\item{psu}{the PSU identifier (integer).}
\item{size}{the number of tracts in the PSU (integer).}
\item{nn_size}{the neighbourhood size to use (integer).}
}
}
\description{
Prepare a PSU object, which can be either a data frame or a list.
}
\details{
If \code{nn_size} is provided through \code{psus}, \code{tracts} and \code{neighbourhood_size} can be ignored.
Otherwise, the neighbourhood size is calculated from the PSU sizes.
The PSU size is derived from the number of \code{tracts} with a PSU id, and neighbourhood sizes scales
linearly with these sizes, starting from \code{neighbourhood_size}.
}
\examples{
prepped_psus = prepare_psus(psus, tracts, 4L);

}
\seealso{
Other prepare:
\code{\link[=prepare_area]{prepare_area()}},
\code{\link[=prepare_categories]{prepare_categories()}},
\code{\link[=prepare_data]{prepare_data()}},
\code{\link[=prepare_tracts]{prepare_tracts()}}
}
\concept{prepare}
