% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_tracts}
\alias{prepare_tracts}
\title{Prepare a tract object}
\usage{
prepare_tracts(tracts)
}
\arguments{
\item{tracts}{Any of the following:
\itemize{
\item A named list with items: \describe{
\item{tract}{the tract identifier (integer).}
\item{psu}{the PSU identifier of the smallest PSU containing the tract (integer).}
\item{auxiliaries}{the auxiliary information of the tracts (float matrix).}
}
\item An unnamed list with items following the item order of the named list.
\item A data frame with items following the item order of the named list. Any columns after the first
two will assumed to be auxiliaries, however discarded if they are non-floaty.
\item A matrix with two columns containing the \code{tract}s and the \code{psu}s, in that order.
}}
}
\value{
A list:
\describe{
\item{tract}{the category identifier (integer).}
\item{psu}{the PSU identifier of the smallest PSU containing the tract (integer).}
\item{auxiliaries}{the auxiliary information of the tracts (float matrix).}
}
}
\description{
Prepare a category object, which can be either a list or a data frame.
}
\examples{
prepped_tracts = prepare_tracts(tracts);

}
\seealso{
Other prepare:
\code{\link[=prepare_area]{prepare_area()}},
\code{\link[=prepare_categories]{prepare_categories()}},
\code{\link[=prepare_data]{prepare_data()}},
\code{\link[=prepare_psus]{prepare_psus()}}
}
\concept{prepare}
