% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nilsier_methods.R
\name{print.nilsier}
\alias{print.nilsier}
\title{Print a NILS estimate}
\usage{
\method{print}{nilsier}(x, complete = TRUE, ...)
}
\arguments{
\item{x}{A \link{nilsier} object.}

\item{complete}{Logical. If \code{FALSE} (default), excludes apparent zero-tracts from the printed
output.}

\item{...}{Additional arguments (currently unused)}
}
\value{
Invisibly returns the input \link{nilsier} object.
}
\description{
Prints a summary of a \link{nilsier} object.
}
