% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nilsier_methods.R
\name{summary.nilsier}
\alias{summary.nilsier}
\alias{print.summary.nilsier}
\title{Summarize a NILS estimate}
\usage{
\method{summary}{nilsier}(object, ...)

\method{print}{summary.nilsier}(x, ...)
}
\arguments{
\item{object}{A \link{nilsier} object.}

\item{...}{Additional arguments (currently unused).}

\item{x}{A \link{summary.nilsier}.}
}
\value{
The returned list has the following components:
\describe{
\item{estimate}{Estimated total of the target variable.}
\item{variance}{Estimated variance of the total estimator.}
\item{rel_se}{Estimated relative standard error of the total estimator.}
\item{positive_tracts}{Number of non-nil tracts. Not shown if filtering has been applied.}
}
}
\description{
Produces summary statistics for a \link{nilsier} object.
}
\examples{
obj = nilsier(psus, categories, tracts, plots);
summary(obj);

}
