% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nilsier_methods.R
\name{coef.nilsier}
\alias{coef.nilsier}
\title{Estimates of a NILS estimate}
\usage{
\method{coef}{nilsier}(obj, ...)
}
\arguments{
\item{...}{Additional arguments (currently unused).}

\item{object}{A \link{nilsier} object.}
}
\value{
A named vector with the following elements:
\describe{
\item{estimate}{Estimated total of the target variable.}
\item{variance}{Estimated variance of the total estimator.}
}
}
\description{
Accesses the estimate and variance estimate of a \link{nilsier} object.
}
\examples{
obj = nilsier(psus, categories, tracts, plots);
coef(obj);

}
