% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nilsier_methods.R
\name{efilter}
\alias{efilter}
\alias{efilter.nilsier}
\title{Filter NILS estimates}
\usage{
efilter(obj, ...)

\method{efilter}{nilsier}(obj, psus = NULL, categories = NULL, ...)
}
\arguments{
\item{obj}{A \link{nilsier} object.}

\item{...}{Additional arguments (currently unused)}

\item{psus}{An optional vector of PSU IDs to keep.}

\item{categories}{An optional vector of the category IDs to keep.}
}
\value{
a filtered \link{nilsier} object.
}
\description{
Filters a \link{nilsier} object by PSU level or category.
}
\examples{
obj = nilsier(psus, categories, tracts, plots);

# Keep only category with ID 1
efilter(obj, categories = 1)

# Keep PSU 5
efilter(obj, psus = 5)


}
