% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nils.R
\name{nilsier}
\alias{nilsier}
\title{Estimate totals using the NILS hierarchical design}
\usage{
nilsier(
  psus,
  categories,
  tracts,
  data,
  frame_area = 46519242.1175867,
  tract_area = 196 * 100 * pi,
  neighbourhood_size = 4L,
  variance_strategy = "srs"
)
}
\arguments{
\item{psus}{A vector of PSU identifiers (integer) in order from the largest to the smallest, or;
a data frame, list, matrix, or vector containing the PSUs:
\describe{
\item{psu}{the PSU identifiers (integer) in order from the largest to the smallest.}
\item{nn_size}{(optional) the neighbourhood size to use (integer > \code{1L}). Will be derived from
\code{neighbourhood_size} if not given.}
}
If the columns (or items) are unnamed, or if a matrix is supplied, the order of the columns
(or items) are assumed to follow the order given by the list above.
If a matrix is supplied, the order is also assumed to be in the parameter list order.}

\item{categories}{A data frame, list, or a matrix containing the categories:
\describe{
\item{category}{the category identifier (integer).}
\item{psu}{the PSU identifier of the smallest PSu in which the category can appear (integer).}
}
If the columns (or items) are unnamed, the order of the columns (or items) are assumed to be
according to the order given by the parameter list.
If a matrix is supplied, the order is also assumed to be in the parameter list order.}

\item{tracts}{Any of the following:
\itemize{
\item A named list with items: \describe{
\item{tract}{the tract identifier (integer).}
\item{psu}{the PSU identifier of the smallest PSU containing the tract (integer).}
\item{auxiliaries}{the auxiliary information of the tracts (float matrix).}
}
\item An unnamed list with items following the item order of the named list.
\item A data frame with items following the item order of the named list. Any columns after the first
two will assumed to be auxiliaries, however discarded if they are non-floaty.
\item A matrix with two columns containing the \code{tract}s and the \code{psu}s, in that order.
}}

\item{data}{A data frame (or list) containing the survey data:
\describe{
\item{tract}{the tract identifier (integer).}
\item{category}{the category identifier (integer).}
\item{dw}{the design weight of the data (float).}
\item{value}{the recorded value (float).}
}
If the columns (or items) are named, the default names can be overidden by the corresponding
argument.
If the columns (or items) are unnamed, the order of the columns (or items) is assumed to be
according to the order given by the parameter list.}

\item{frame_area}{The size of the area frame. Typically larger than the actual area of interest.}

\item{tract_area}{The area of a tract, expressed in the same units as the target variable.}

\item{variance_strategy}{The strategy for estimating variance. One of \code{"srs"} or
\code{"nearest_neighbour"}.}
}
\value{
A \code{nilsier} object, essentially a data frame with one row per category and the
following columns:
\describe{
\item{Cat. ID}{The category ID number.}
\item{Est. total}{The estimated total of the target variable within the category.}
\item{Est. variance}{The estimated variance of the total estimator within the category.}
\item{Positive tracts}{The number of tracts with at least one positive value of the
target variable in the category.}
}
}
\description{
Estimates the total of some variable surveyed under the NILS hierarchical sampling framework.
}
\details{
The function combines plot-level observations (\code{plot_data}), tract-level information
(\code{tract_data}), PSU hierarchy (\code{psus}), and category assignments (\code{category_psu_map}) to estimate
totals under the NILS sampling design.
\subsection{Variance estimation strategy}{

If the \code{psus} were selected using a spatially balanced sampling design, the \code{"srs"} variance
strategy often overestimates the variance.

When using the \code{"nearest_neighbours"} strategy, a local neighbourhood variance is used.
The neighbourhood size is either manually supplied or calculated from the psu sizes (see
\code{\link[=prepare_psus]{prepare_psus()}}). When calculated from the sizes, the PSUs scales linearly with the nearest
neighbourhood size of the smallest PSU, given by \code{neighbourhood_size}.
}
}
\examples{
obj = nilsier(psus, category_psu_map, tracts, plots);

}
