% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_categories}
\alias{prepare_categories}
\title{Prepare a category object}
\usage{
prepare_categories(categories)
}
\arguments{
\item{categories}{A data frame, list, or a matrix containing the categories:
\describe{
\item{category}{the category identifier (integer).}
\item{psu}{the PSU identifier of the smallest PSu in which the category can appear (integer).}
}
If the columns (or items) are unnamed, the order of the columns (or items) is assumed to be
according to the order given by the parameter list.
If a matrix is supplied, the order is also assumed to be in the parameter list order.}
}
\value{
A data frame
\describe{
\item{category}{the category identifier (integer).}
\item{psu}{the PSU identifier of the smallest PSU in which the category can appear (integer).}
}
}
\description{
Prepare a category object, which can be either a data frame or a list.
}
\examples{
prepped_cats = prepare_categories(categories);

}
\seealso{
Other prepare:
\code{\link[=prepare_area]{prepare_area()}},
\code{\link[=prepare_data]{prepare_data()}},
\code{\link[=prepare_psus]{prepare_psus()}},
\code{\link[=prepare_tracts]{prepare_tracts()}}
}
\concept{prepare}
