% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nilsier_methods.R
\name{vcov.nilsier}
\alias{vcov.nilsier}
\title{Covariance matrix for NILS category estimates}
\usage{
\method{vcov}{nilsier}(obj, complete = TRUE, ...)
}
\arguments{
\item{complete}{Logical. If \code{FALSE}, excludes apparent zero-tracts.}

\item{...}{Additional arguments (currently unused).}

\item{object}{A \link{nilsier} object.}
}
\value{
the covariance matrix of the \link{nilsier} object.
}
\description{
Accesses the covariance matrix of a \link{nilsier} object.
}
\examples{
obj = NilsEstimate(plots, tracts, psus, category_psu_map);
vcov(obj);

}
